Update README.md
The objective of this project is to determine the efficacy of various deep learning methods for accurately classifying images acquired by the Mars Curiosity Rover. Through this study, we intend to identify the advantages, disadvantages, and potential synergies of these three image classification techniques, thereby providing a comprehensive guide for optimizing image classification for Mars Rover missions.